+++ packages/openapi3/src/openapi.ts
@@ -975,7 +975,7 @@ function createOAPIEmitter(
   ) {
     schemaMap ??= new Map<string, OpenAPI3MediaType[]>();
     for (const data of responses) {
-      if (data.body === undefined) {
+      if (data.body === undefined || isVoidType(data.body.type)) {
         continue;
       }
       obj.content ??= {};