+++ packages/openapi3/test/return-types.test.ts
@@ -76,6 +76,23 @@ worksFor(["3.0.0", "3.1.0"], ({ checkFor, openApiFor }) => {
     ok(res.paths["/"].put.responses["201"].content["application/json"].schema);
   });
 
+  it("handle @statusCode decorator with void return type", async () => {
+    const res = await openApiFor(`
+      model TestResult<T> {
+        @statusCode statusCode: 201;
+        @body body: T;
+      }
+
+      interface Test {
+        test(): TestResult<void>;
+      }
+      `);
+    ok(res.paths["/"].get.responses["201"]);
+    deepStrictEqual(res.paths["/"].get.responses["201"], {
+      description: "The request has succeeded and a new resource has been created as a result.",
+    });
+  });
+
   it("defines responses with headers and status codes", async () => {
     const res = await openApiFor(
       `