+++ packages/http-client-python/CHANGELOG.md
@@ -1,5 +1,35 @@
 # Change Log - @typespec/http-client-python
 
+## 0.6.1
+
+### Bug Fixes
+
+- Generate `__init__` for all DPG models to ensure stability
+
+## 0.6.0
+
+### Features
+
+- Add support for typespec namespace
+
+## 0.5.1
+
+### Bug Fixes
+
+- Do not do exception sort if there is no operation groups
+
+## 0.5.0
+
+### Features
+
+- Add support for generation in enviroments without a Python installation
+
+## 0.4.4
+
+### Bug Fixes
+
+- `:code:` in docstring should always be preceded by `\`
+
 ## 0.4.3
 
 ### Bump dependencies